pyproject.toml
[project]
name = "calendar-prioritizer"
version = "0.1.0"
description = "FastAPI backend for Google Calendar OAuth and calendar retrieval."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi",
    "google-api-python-client",
    "google-auth",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "httpx",
    "itsdangerous",
    "pydantic-settings",
    "python-dotenv",
    "sqlalchemy",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "coverage",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/calendar_prioritizer"]

[tool.uv.workspace]
members = [
    "src/calendar_prioritizer/tests",
]
